Fix adc read to be able to support 32 bit resolution adcs

+++ common/drivers/platform/stm32f4/st_adc.cc
@@ -162,7 +162,7 @@ bool HwAdc::en_dma_req()
     return false;
 }
 
-bool HwAdc::read(uint16_t& val)
+bool HwAdc::read(uint32_t& val)
 {
     // Wait until ADC conversion is finished
     while ((base_addr->SR & ADC_SR_EOC) == 0);
